% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats_calibration.R
\name{calibration_summary_stats}
\alias{calibration_summary_stats}
\title{Estimate the correlation between external and internal trio gwas (i.e. sample overlap)}
\usage{
calibration_summary_stats(df, rho, family = "trio", r = "shared")
}
\arguments{
\item{df:}{a data frame with these columns: "SNP", "beta_int", "beta_int_var", "alpha_int", "alpha_int_var",
"alpha_ext", "alpha_ext_var", "N_int", "N_ext"}

\item{rho:}{sample overlap ratio}

\item{family:}{type of family structure. Either trio or sibling}

\item{r:}{assumption on correlation between T and NT. Can be zero, shared, or a provided vector.}
}
\value{
a data frame of snp id, calibrated beta and its standard error
}
\description{
Estimate the correlation between external and internal trio gwas (i.e. sample overlap)
}
