% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibling_calibration.R
\name{format_data}
\alias{format_data}
\title{Format the data for calibration (shared across SNPs)}
\usage{
format_data(pheno, F_ind, Z = NULL)
}
\arguments{
\item{pheno:}{N-vector of individual phenotype}

\item{F_ind:}{N-vector of invidual family index}

\item{Z:}{N by p data frame of covariates}
}
\value{
a list of following quantities
(1) N: number of individuals
(2) K: number of families
(3) Y: N-vector of phenotype
(4) Y_tilde: N vector of normalized phenotype
(5) Z: N by p data frame of covariates
(6) Z_tilde: N by p data frame of normalized covariates
(7) F_ind: N vector of family index
(8) size_dic: K vector of family size of each family
(9) F_size: N vector of family size of each individual
(10) dim_Z: dimension of Z
(11) family2ind: a K-list such that the kth element contains indices of
indicies from family k
}
\description{
Format the data for calibration (shared across SNPs)
}
