% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats_calibration.R
\name{sample_overlap}
\alias{sample_overlap}
\title{Estimate the correlation between external and internal gwas (i.e. sample overlap)}
\usage{
sample_overlap(df, snp_list = NA)
}
\arguments{
\item{df:}{a data frame with these columns: "SNP", "beta_int", "beta_int_var", "alpha_int", "alpha_int_var",
"alpha_ext", "alpha_ext_var", "N_int", "N_ext"}

\item{snp_list:}{list of SNPs that has zero direct effects (can be selected from an independent selection file)}
}
\value{
rho: sample overlap ratio
}
\description{
Estimate the correlation between external and internal gwas (i.e. sample overlap)
}
